+++ android/app/src/main/java/com/woowacourse/friendogly/data/model/FootprintDto.kt
@@ -3,7 +3,7 @@ package com.woowacourse.friendogly.data.model
 import kotlinx.datetime.LocalDateTime
 
 data class FootprintDto(
-    val footPrintId: Long,
+    val footprintId: Long,
     val latitude: Double,
     val longitude: Double,
     val createdAt: LocalDateTime,
