+++ android/app/src/main/java/com/woowacourse/friendogly/presentation/ui/MainActivity.kt
@@ -25,6 +25,7 @@ class MainActivity :
 
     override fun initCreateView() {
         initNavController()
+        requestLocationPermissions()
     }
 
     private fun initNavController() {
