+++ android/app/src/main/java/com/woowacourse/friendogly/presentation/ui/woof/WoofFragment.kt
@@ -1,5 +1,7 @@
 package com.woowacourse.friendogly.presentation.ui.woof
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import android.provider.Settings
 import android.view.View
