+++ android/app/src/main/java/com/woowacourse/friendogly/remote/mapper/FootprintInfoMapper.kt
@@ -1,16 +1,16 @@
-package com.woowacourse.friendogly.data.mapper
+package com.woowacourse.friendogly.remote.mapper
 
 import com.woowacourse.friendogly.data.model.FootprintInfoDto
-import com.woowacourse.friendogly.domain.model.FootprintInfo
+import com.woowacourse.friendogly.remote.model.response.FootprintInfoResponse
 
-fun FootprintInfoDto.toDomain(): FootprintInfo {
-    return FootprintInfo(
+fun FootprintInfoResponse.toData(): FootprintInfoDto {
+    return FootprintInfoDto(
         memberName = memberName,
         petName = petName,
         petDescription = petDescription,
         petBirthDate = petBirthDate,
-        petSizeType = petSizeType.toDomain(),
-        petGender = petGender.toDomain(),
+        petSizeType = petSizeType.toData(),
+        petGender = petGender.toData(),
         footprintImageUrl = footprintImageUrl,
         createdAt = createdAt,
         isMine = isMine,
