+++ android/app/src/main/java/com/woowacourse/friendogly/remote/source/WoofDataSourceImpl.kt
@@ -1,12 +1,12 @@
 package com.woowacourse.friendogly.remote.source
 
-import com.woowacourse.friendogly.data.mapper.toData
 import com.woowacourse.friendogly.data.model.FootprintDto
 import com.woowacourse.friendogly.data.model.FootprintMarkBtnInfoDto
 import com.woowacourse.friendogly.data.model.FootprintSaveDto
 import com.woowacourse.friendogly.data.model.LandMarkDto
 import com.woowacourse.friendogly.data.source.WoofDataSource
 import com.woowacourse.friendogly.remote.api.WoofService
+import com.woowacourse.friendogly.remote.mapper.toData
 import com.woowacourse.friendogly.remote.model.request.FootprintRequest
 
 class WoofDataSourceImpl(private val service: WoofService) : WoofDataSource {
